Add an automatic test script
Testscript compares testcase behavior between notepad and vim, to ensure the script duplicates behaviour as best as it can.

The script can be temperamental, depending on the computer it's being run on. Increasing the `setkeydelay` at the top of the script does help, but slows down the tests.

Currently has a few testcases, a few of which fail because this script does have bugs or differences to vim's behaviour.

Adapted from [my script here](https://github.com/idvorkin/Vim-Keybindings-For-Onenote/)